Added support for Project Administrator username and password.  This allows the remote Administrator to execute SQL commands on a running Project.  In the future, this may allow more freedom to administrate the Server/Project.
Added new index to Test.  The ChartServlet query to build the timing
chart was not performing well on a large-ish database, with the new
index, the charts appear _much_ faster.  This is aided by the change
above.

Added getSchedulerStatus call into the Server and DartClient.  New the
scheduler status can be queried independantly of the rest of the Dart
Server.  We hope to be able to better debug hung servers.


git-svn-id: http://svn.na-mic.org/Dart/trunk@356 b2c14c61-d7ef-0310-8715-841e9ab387fe